Streamlined interface for clients to interact with chatbots.
Will make it easier to enable multiple different chatbot architectures.
